Remove all wp.api Usage from the editor package
