+++ editor/store/effects.js
@@ -1,8 +1,7 @@
 /**
  * External dependencies
  */
-import { BEGIN, COMMIT, REVERT } from 'redux-optimist';
-import { get, includes, last, map, castArray, uniqueId, pick } from 'lodash';
+import { get, last } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -12,70 +11,55 @@ import {
 	getBlockType,
 	switchToBlockType,
 	createBlock,
-	serialize,
-	isSharedBlock,
 	getDefaultBlockForPostFormat,
 	doBlocksMatchTemplate,
 	synchronizeBlocksWithTemplate,
 } from '@wordpress/blocks';
 import { __ } from '@wordpress/i18n';
 import { speak } from '@wordpress/a11y';
-import apiFetch from '@wordpress/api-fetch';
 
 /**
  * Internal dependencies
  */
 import {
 	setupEditorState,
-	resetAutosave,
-	resetPost,
-	updatePost,
-	receiveBlocks,
-	receiveSharedBlocks,
-	replaceBlock,
 	replaceBlocks,
-	createSuccessNotice,
-	createErrorNotice,
 	createWarningNotice,
-	removeNotice,
-	saveSharedBlock,
 	insertBlock,
-	removeBlocks,
 	selectBlock,
 	removeBlock,
 	resetBlocks,
 	setTemplateValidity,
 } from './actions';
 import {
-	getCurrentPost,
-	getCurrentPostType,
-	getEditedPostContent,
-	getPostEdits,
-	isEditedPostAutosaveable,
-	isEditedPostSaveable,
 	getBlock,
 	getBlockCount,
 	getBlockRootClientId,
 	getBlocks,
-	getSharedBlock,
 	getPreviousBlockClientId,
 	getProvisionalBlockClientId,
 	getSelectedBlock,
 	isBlockSelected,
 	getTemplate,
 	getTemplateLock,
-	getAutosave,
-	isEditedPostNew,
-	POST_UPDATE_TRANSACTION_ID,
 } from './selectors';
-
-/**
- * Module Constants
- */
-const SAVE_POST_NOTICE_ID = 'SAVE_POST_NOTICE_ID';
-const AUTOSAVE_POST_NOTICE_ID = 'AUTOSAVE_POST_NOTICE_ID';
-const TRASH_POST_NOTICE_ID = 'TRASH_POST_NOTICE_ID';
-const SHARED_BLOCK_NOTICE_ID = 'SHARED_BLOCK_NOTICE_ID';
+import {
+	fetchSharedBlocks,
+	saveSharedBlocks,
+	deleteSharedBlocks,
+	convertBlockToShared,
+	convertBlockToStatic,
+	receiveSharedBlocks,
+} from './effects/shared-blocks';
+import {
+	requestPostUpdate,
+	requestPostUpdateSuccess,
+	requestPostUpdateFailure,
+	trashPost,
+	trashPostFailure,
+	refreshPost,
+	AUTOSAVE_POST_NOTICE_ID,
+} from './effects/posts';
 
 /**
  * Effect handler returning an action to remove the provisional block, if one
@@ -95,239 +79,17 @@ export function removeProvisionalBlock( action, store ) {
 }
 
 export default {
-	REQUEST_POST_UPDATE( action, store ) {
-		const { dispatch, getState } = store;
-		const state = getState();
-		const post = getCurrentPost( state );
-		const isAutosave = !! action.options.autosave;
-
-		// Prevent save if not saveable.
-		const isSaveable = isAutosave ? isEditedPostAutosaveable : isEditedPostSaveable;
-		if ( ! isSaveable( state ) ) {
-			return;
-		}
-
-		let edits = getPostEdits( state );
-		if ( isAutosave ) {
-			edits = pick( edits, [ 'title', 'content', 'excerpt' ] );
-		}
-
-		// New posts (with auto-draft status) must be explicitly assigned draft
-		// status if there is not already a status assigned in edits (publish).
-		// Otherwise, they are wrongly left as auto-draft. Status is not always
-		// respected for autosaves, so it cannot simply be included in the pick
-		// above. This behavior relies on an assumption that an auto-draft post
-		// would never be saved by anyone other than the owner of the post, per
-		// logic within autosaves REST controller to save status field only for
-		// draft/auto-draft by current user.
-		//
-		// See: https://core.trac.wordpress.org/ticket/43316#comment:88
-		// See: https://core.trac.wordpress.org/ticket/43316#comment:89
-		if ( isEditedPostNew( state ) ) {
-			edits = { status: 'draft', ...edits };
-		}
-
-		let toSend = {
-			...edits,
-			content: getEditedPostContent( state ),
-			id: post.id,
-		};
-		const basePath = wp.api.getPostTypeRoute( getCurrentPostType( state ) );
-
-		dispatch( {
-			type: 'REQUEST_POST_UPDATE_START',
-			optimist: { type: BEGIN, id: POST_UPDATE_TRANSACTION_ID },
-			isAutosave,
-		} );
-
-		// Optimistically apply updates under the assumption that the post
-		// will be updated. See below logic in success resolution for revert
-		// if the autosave is applied as a revision.
-		dispatch( {
-			...updatePost( toSend ),
-			optimist: { id: POST_UPDATE_TRANSACTION_ID },
-		} );
-
-		let request;
-		if ( isAutosave ) {
-			// Ensure autosaves contain all expected fields, using autosave or
-			// post values as fallback if not otherwise included in edits.
-			toSend = {
-				...pick( post, [ 'title', 'content', 'excerpt' ] ),
-				...getAutosave( state ),
-				...toSend,
-				parent: post.id,
-			};
-
-			request = apiFetch( {
-				path: `/wp/v2/${ basePath }/${ post.id }/autosaves`,
-				method: 'POST',
-				data: toSend,
-			} );
-		} else {
-			dispatch( removeNotice( SAVE_POST_NOTICE_ID ) );
-			dispatch( removeNotice( AUTOSAVE_POST_NOTICE_ID ) );
-
-			request = apiFetch( {
-				path: `/wp/v2/${ basePath }/${ post.id }`,
-				method: 'PUT',
-				data: toSend,
-			} );
-		}
-
-		request
-			.then( ( newPost ) => {
-				const reset = isAutosave ? resetAutosave : resetPost;
-				dispatch( reset( newPost ) );
-
-				// An autosave may be processed by the server as a regular save
-				// when its update is requested by the author and the post was
-				// draft or auto-draft.
-				const isRevision = newPost.id !== post.id;
-
-				dispatch( {
-					type: 'REQUEST_POST_UPDATE_SUCCESS',
-					previousPost: post,
-					post: newPost,
-					optimist: {
-						// Note: REVERT is not a failure case here. Rather, it
-						// is simply reversing the assumption that the updates
-						// were applied to the post proper, such that the post
-						// treated as having unsaved changes.
-						type: isRevision ? REVERT : COMMIT,
-						id: POST_UPDATE_TRANSACTION_ID,
-					},
-					isAutosave,
-				} );
-			} )
-			.catch( ( error ) => dispatch( {
-				type: 'REQUEST_POST_UPDATE_FAILURE',
-				optimist: { type: REVERT, id: POST_UPDATE_TRANSACTION_ID },
-				post,
-				edits,
-				error,
-			} ) );
+	REQUEST_POST_UPDATE: ( action, store ) => {
+		requestPostUpdate( action, store );
 	},
-	REQUEST_POST_UPDATE_SUCCESS( action, store ) {
-		const { previousPost, post, isAutosave } = action;
-		const { dispatch, getState } = store;
-
-		// TEMPORARY: If edits remain after a save completes, the user must be
-		// prompted about unsaved changes. This should be refactored as part of
-		// the `isEditedPostDirty` selector instead.
-		//
-		// See: https://github.com/WordPress/gutenberg/issues/7409
-		if ( Object.keys( getPostEdits( getState() ) ).length ) {
-			dispatch( { type: 'DIRTY_ARTIFICIALLY' } );
-		}
-
-		// Autosaves are neither shown a notice nor redirected.
-		if ( isAutosave ) {
-			return;
-		}
-
-		const publishStatus = [ 'publish', 'private', 'future' ];
-		const isPublished = includes( publishStatus, previousPost.status );
-		const willPublish = includes( publishStatus, post.status );
-
-		let noticeMessage;
-		let shouldShowLink = true;
-
-		if ( ! isPublished && ! willPublish ) {
-			// If saving a non-published post, don't show notice.
-			noticeMessage = null;
-		} else if ( isPublished && ! willPublish ) {
-			// If undoing publish status, show specific notice
-			noticeMessage = __( 'Post reverted to draft.' );
-			shouldShowLink = false;
-		} else if ( ! isPublished && willPublish ) {
-			// If publishing or scheduling a post, show the corresponding
-			// publish message
-			noticeMessage = {
-				publish: __( 'Post published!' ),
-				private: __( 'Post published privately!' ),
-				future: __( 'Post scheduled!' ),
-			}[ post.status ];
-		} else {
-			// Generic fallback notice
-			noticeMessage = __( 'Post updated!' );
-		}
-
-		if ( noticeMessage ) {
-			dispatch( createSuccessNotice(
-				<p>
-					{ noticeMessage }
-					{ ' ' }
-					{ shouldShowLink && <a href={ post.link }>{ __( 'View post' ) }</a> }
-				</p>,
-				{ id: SAVE_POST_NOTICE_ID, spokenMessage: noticeMessage }
-			) );
-		}
+	REQUEST_POST_UPDATE_SUCCESS: requestPostUpdateSuccess,
+	REQUEST_POST_UPDATE_FAILURE: requestPostUpdateFailure,
+	TRASH_POST: ( action, store ) => {
+		trashPost( action, store );
 	},
-	REQUEST_POST_UPDATE_FAILURE( action, store ) {
-		const { post, edits, error } = action;
-
-		if ( error && 'rest_autosave_no_changes' === error.code ) {
-			// Autosave requested a new autosave, but there were no changes. This shouldn't
-			// result in an error notice for the user.
-			return;
-		}
-
-		const { dispatch } = store;
-
-		const publishStatus = [ 'publish', 'private', 'future' ];
-		const isPublished = publishStatus.indexOf( post.status ) !== -1;
-		// If the post was being published, we show the corresponding publish error message
-		// Unless we publish an "updating failed" message
-		const messages = {
-			publish: __( 'Publishing failed' ),
-			private: __( 'Publishing failed' ),
-			future: __( 'Scheduling failed' ),
-		};
-		const noticeMessage = ! isPublished && publishStatus.indexOf( edits.status ) !== -1 ?
-			messages[ edits.status ] :
-			__( 'Updating failed' );
-		dispatch( createErrorNotice( noticeMessage, { id: SAVE_POST_NOTICE_ID } ) );
-	},
-	TRASH_POST( action, store ) {
-		const { dispatch, getState } = store;
-		const { postId } = action;
-		const basePath = wp.api.getPostTypeRoute( getCurrentPostType( getState() ) );
-		dispatch( removeNotice( TRASH_POST_NOTICE_ID ) );
-		apiFetch( { path: `/wp/v2/${ basePath }/${ postId }`, method: 'DELETE' } )
-			.then( () => {
-				const post = getCurrentPost( getState() );
-
-				// TODO: This should be an updatePost action (updating subsets of post properties),
-				// But right now editPost is tied with change detection.
-				dispatch( resetPost( { ...post, status: 'trash' } ) );
-			} )
-			.catch( ( error ) => dispatch( {
-				...action,
-				type: 'TRASH_POST_FAILURE',
-				error,
-			} ) );
-	},
-	TRASH_POST_FAILURE( action, store ) {
-		const message = action.error.message && action.error.code !== 'unknown_error' ? action.error.message : __( 'Trashing failed' );
-		store.dispatch( createErrorNotice( message, { id: TRASH_POST_NOTICE_ID } ) );
-	},
-	REFRESH_POST( action, store ) {
-		const { dispatch, getState } = store;
-
-		const state = getState();
-		const post = getCurrentPost( state );
-		const basePath = wp.api.getPostTypeRoute( getCurrentPostType( state ) );
-
-		const data = {
-			context: 'edit',
-		};
-
-		apiFetch( { path: `/wp/v2/${ basePath }/${ post.id }`, data } ).then(
-			( newPost ) => {
-				dispatch( resetPost( newPost ) );
-			}
-		);
+	TRASH_POST_FAILURE: trashPostFailure,
+	REFRESH_POST: ( action, store ) => {
+		refreshPost( action, store );
 	},
 	MERGE_BLOCKS( action, store ) {
 		const { dispatch } = store;
@@ -454,180 +216,18 @@ export default {
 
 		return setTemplateValidity( isValid );
 	},
-	FETCH_SHARED_BLOCKS( action, store ) {
-		// TODO: these are potentially undefined, this fix is in place
-		// until there is a filter to not use shared blocks if undefined
-		const basePath = wp.api.getPostTypeRoute( 'wp_block' );
-		if ( ! basePath ) {
-			return;
-		}
-
-		const { id } = action;
-		const { dispatch } = store;
-
-		let result;
-		if ( id ) {
-			result = apiFetch( { path: `/wp/v2/${ basePath }/${ id }` } );
-		} else {
-			result = apiFetch( { path: `/wp/v2/${ basePath }?per_page=-1` } );
-		}
-
-		result
-			.then( ( sharedBlockOrBlocks ) => {
-				dispatch( receiveSharedBlocks( map(
-					castArray( sharedBlockOrBlocks ),
-					( sharedBlock ) => ( {
-						sharedBlock,
-						parsedBlock: parse( sharedBlock.content )[ 0 ],
-					} )
-				) ) );
-
-				dispatch( {
-					type: 'FETCH_SHARED_BLOCKS_SUCCESS',
-					id,
-				} );
-			} )
-			.catch( ( error ) => dispatch( {
-				type: 'FETCH_SHARED_BLOCKS_FAILURE',
-				id,
-				error,
-			} ) );
+	FETCH_SHARED_BLOCKS: ( action, store ) => {
+		fetchSharedBlocks( action, store );
 	},
-	RECEIVE_SHARED_BLOCKS( action ) {
-		return receiveBlocks( map( action.results, 'parsedBlock' ) );
+	SAVE_SHARED_BLOCK: ( action, store ) => {
+		saveSharedBlocks( action, store );
 	},
-	SAVE_SHARED_BLOCK( action, store ) {
-		// TODO: these are potentially undefined, this fix is in place
-		// until there is a filter to not use shared blocks if undefined
-		const basePath = wp.api.getPostTypeRoute( 'wp_block' );
-		if ( ! basePath ) {
-			return;
-		}
-
-		const { id } = action;
-		const { dispatch } = store;
-		const state = store.getState();
-
-		const { clientId, title, isTemporary } = getSharedBlock( state, id );
-		const { name, attributes, innerBlocks } = getBlock( state, clientId );
-		const content = serialize( createBlock( name, attributes, innerBlocks ) );
-
-		const data = isTemporary ? { title, content } : { id, title, content };
-		const path = isTemporary ? `/wp/v2/${ basePath }` : `/wp/v2/${ basePath }/${ id }`;
-		const method = isTemporary ? 'POST' : 'PUT';
-
-		apiFetch( { path, data, method } )
-			.then( ( updatedSharedBlock ) => {
-				dispatch( {
-					type: 'SAVE_SHARED_BLOCK_SUCCESS',
-					updatedId: updatedSharedBlock.id,
-					id,
-				} );
-				const message = isTemporary ? __( 'Block created.' ) : __( 'Block updated.' );
-				dispatch( createSuccessNotice( message, { id: SHARED_BLOCK_NOTICE_ID } ) );
-			} )
-			.catch( ( error ) => {
-				dispatch( { type: 'SAVE_SHARED_BLOCK_FAILURE', id } );
-				dispatch( createErrorNotice( error.message, {
-					id: SHARED_BLOCK_NOTICE_ID,
-					spokenMessage: error.message,
-				} ) );
-			} );
+	DELETE_SHARED_BLOCK: ( action, store ) => {
+		deleteSharedBlocks( action, store );
 	},
-	DELETE_SHARED_BLOCK( action, store ) {
-		// TODO: these are potentially undefined, this fix is in place
-		// until there is a filter to not use shared blocks if undefined
-		const basePath = wp.api.getPostTypeRoute( 'wp_block' );
-		if ( ! basePath ) {
-			return;
-		}
-
-		const { id } = action;
-		const { getState, dispatch } = store;
-
-		// Don't allow a shared block with a temporary ID to be deleted
-		const sharedBlock = getSharedBlock( getState(), id );
-		if ( ! sharedBlock || sharedBlock.isTemporary ) {
-			return;
-		}
-
-		// Remove any other blocks that reference this shared block
-		const allBlocks = getBlocks( getState() );
-		const associatedBlocks = allBlocks.filter( ( block ) => isSharedBlock( block ) && block.attributes.ref === id );
-		const associatedBlockClientIds = associatedBlocks.map( ( block ) => block.clientId );
-
-		const transactionId = uniqueId();
-
-		dispatch( {
-			type: 'REMOVE_SHARED_BLOCK',
-			id,
-			optimist: { type: BEGIN, id: transactionId },
-		} );
-
-		// Remove the parsed block.
-		dispatch( removeBlocks( [
-			...associatedBlockClientIds,
-			sharedBlock.clientId,
-		] ) );
-
-		apiFetch( { path: `/wp/v2/${ basePath }/${ id }`, method: 'DELETE' } )
-			.then( () => {
-				dispatch( {
-					type: 'DELETE_SHARED_BLOCK_SUCCESS',
-					id,
-					optimist: { type: COMMIT, id: transactionId },
-				} );
-				const message = __( 'Block deleted.' );
-				dispatch( createSuccessNotice( message, { id: SHARED_BLOCK_NOTICE_ID } ) );
-			} )
-			.catch( ( error ) => {
-				dispatch( {
-					type: 'DELETE_SHARED_BLOCK_FAILURE',
-					id,
-					optimist: { type: REVERT, id: transactionId },
-				} );
-				dispatch( createErrorNotice( error.message, {
-					id: SHARED_BLOCK_NOTICE_ID,
-					spokenMessage: error.message,
-				} ) );
-			} );
-	},
-	CONVERT_BLOCK_TO_STATIC( action, store ) {
-		const state = store.getState();
-		const oldBlock = getBlock( state, action.clientId );
-		const sharedBlock = getSharedBlock( state, oldBlock.attributes.ref );
-		const referencedBlock = getBlock( state, sharedBlock.clientId );
-		const newBlock = createBlock( referencedBlock.name, referencedBlock.attributes );
-		store.dispatch( replaceBlock( oldBlock.clientId, newBlock ) );
-	},
-	CONVERT_BLOCK_TO_SHARED( action, store ) {
-		const { getState, dispatch } = store;
-
-		const parsedBlock = getBlock( getState(), action.clientId );
-		const sharedBlock = {
-			id: uniqueId( 'shared' ),
-			clientId: parsedBlock.clientId,
-			title: __( 'Untitled shared block' ),
-		};
-
-		dispatch( receiveSharedBlocks( [ {
-			sharedBlock,
-			parsedBlock,
-		} ] ) );
-
-		dispatch( saveSharedBlock( sharedBlock.id ) );
-
-		dispatch( replaceBlock(
-			parsedBlock.clientId,
-			createBlock( 'core/block', {
-				ref: sharedBlock.id,
-				layout: parsedBlock.attributes.layout,
-			} )
-		) );
-
-		// Re-add the original block to the store, since replaceBlock() will have removed it
-		dispatch( receiveBlocks( [ parsedBlock ] ) );
-	},
+	RECEIVE_SHARED_BLOCKS: receiveSharedBlocks,
+	CONVERT_BLOCK_TO_STATIC: convertBlockToStatic,
+	CONVERT_BLOCK_TO_SHARED: convertBlockToShared,
 	CREATE_NOTICE( { notice: { content, spokenMessage } } ) {
 		const message = spokenMessage || content;
 		speak( message, 'assertive' );
