+++ packages/data/src/registry.js
@@ -182,6 +182,7 @@ export function createRegistry( storeConfigs = {} ) {
 				// Attempt to normalize fulfillment as async iterable.
 				fulfillment = toAsyncIterable( fulfillment );
 				if ( ! isAsyncIterable( fulfillment ) ) {
+					finishResolution( reducerKey, selectorName, args );
 					return;
 				}
